Fix hand counts in switcher z fighting when playing card to pile
When playing a card it would go below the card counts of the switcher.
With this commit you can set the zIndex of Count and Card to finetune
the z positioning.
Also in some cases, refactored z positioning by using the order
of HTML elements instead of using a parent container with zIndex set.